feat: Refactored table component
