+++ splunk_add_on_ucc_framework/ucc_ui_lib/src/main/webapp/pages/Configuration/ConfigurationPage.jsx
@@ -1,0 +1,14 @@
+import React from 'react';
+import { TitleComponent, SubTitleComponent } from '../Input/InputPageStyle';
+
+function ConfigurationPage() {
+    return (
+        <>
+            <TitleComponent>Configuration</TitleComponent>
+            <SubTitleComponent>Set up your add-on</SubTitleComponent>
+            <hr />
+        </>
+    )
+};
+
+export default ConfigurationPage;
