+++ splunk_add_on_ucc_framework/ucc_ui_lib/src/main/webapp/util/util.js
